ml-training/requirements.txt
# ML Training Dependencies
# Install with: pip install -r requirements.txt

# Core ML/DL frameworks
tensorflow>=2.13.0
keras>=2.13.0
torch>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Firebase
firebase-admin>=6.2.0

# Model persistence
joblib>=1.3.0

# Progress bars
tqdm>=4.65.0

# Jupyter notebooks (optional)
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
